requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
fpdf==1.7.2
idna==3.4
Pillow==9.3.0
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
